Move the button assigner and motion logic to the Ryujinx.Input project
