+++ Ryujinx.Input/Motion/CemuHook/Client.cs
