+++ Ryujinx.Input/Motion/CemuHook/Protocol/ControllerData.cs
