+++ Ryujinx.Input/Motion/CemuHook/Protocol/ControllerInfo.cs
