+++ Ryujinx.Input/Motion/CemuHook/Protocol/Header.cs
