+++ Ryujinx.Input/Motion/CemuHook/Protocol/MessageType.cs
