+++ Ryujinx.Input/Motion/CemuHook/Protocol/SharedResponse.cs
